symbol code: improve EnsureReadable for Win32
This is necessary because we may try to read a block of code that is no
longer mapped in memory (e.g. JITs or recently unloaded DLLs).

To work around memcpy() faulting on us, we can just use
ReadProcessMemory of 1 byte on the address, and ensure that it's
readable that way.

This does avoid crashes most of the time, but a race condition is
possible:

 - We check EnsureReadable, it succeeds
 - Another thread unloads a DLL with the region we wanted to access
 - We attempt the memcpy, and crash

One way around this is to use the LdrLockLoaderLock/LdrUnlockLoaderLock
APIs to ensure the region stays readable until we finish.

Unfortunately none of this is really portable, it'd be nice if there was
a better non-faulting read method.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>